# pg_dump: Rework binary upgrade metadata queries.