This PR refactors the functions used to gather metadata required for binary upgrade schema dumps to avoid repetitive queries in performance-critical paths. 

Upgrades require us to preassign the OIDs of tables, indexes, and types so that they match in the old and new clusters. Previously, multiple queries were executed for each object to gather the necessary info to record in the dump file. Now we gather all of the info required up front and use hash table lookups to set and dump the correct OIDs.

Currently, a database with 100k AO tables can result in well over 1 million queries just to gather the required metadata from the old cluster. This PR eliminates those queries.

The names of pg_toast, pg_aoseg, and pg_bitmapindex tables for the given table/index are set to match what they will be in the new cluster, using the OID of the owning table. In many cases these will be the same as the ones in the old cluster, but not always. Some operations can cause the old table names not to match its owner's OID, but the new cluster will be using the correct name, and it's the new cluster's name that we have to use

The AOTableInfo and BMIndxInfo structs have been added, and metadata fields have been added to the end of the TableInfo, IndxInfo, and TypeInfo structs.

Also, pg_dump now locks all interesting tables in a single statement. This reduces the potentially significant overhead of dispatching many LOCK TABLE queries and reduces the window where external DDL changes can occur.

The query to check for `DISTRIBUTED BY` is now embedded in the single getTables query instead of executing once per table.

These changes result in a significant performance improvement in all cases, but particularly when handling AO tables.

A dump of the regression database shows a performance increase of over 20x and continues to scale as object count increases.

Old: 
```
time pg_dump --schema-only --quote-all-identifiers -s regression --binary-upgrade      
0.82s user 0.21s system 1% cpu 1:27.17 total
```

PR:
```
time pg_dump --schema-only --quote-all-identifiers -s regression --binary-upgrade
0.43s user 0.06s system 12% cpu 3.966 total
```

To illustrate the reason why AO tables are a limiting factor, currently the creation of an AO table actually creates a minimum of 7 catalog objects. When doing a binary upgrade schema dump, we need to record the OIDs of each of these 7 objects to restore in the new cluster. Currently, that requires 7 queries to the old database to collect the info for 1 AO table. 
Practically, most AO tables have an index and a toast table, meaning we need to create 13 objects. 
This gets out of hand very quickly if the AO table is heavily partitioned and this doesn't consider the pg_upgrade overhead with AO tables, only pg_dump. Currently, every type object being dumped does an extra query checking for typarray value, which is always 0 for these composite types, but the query is sent anyway, so we add 5 more queries for a total of 18 for the AO table.

Here's the OID assignment for an AO table with a toast field and an index.

```
"adding OID assignment: catalog ""1259"", namespace: 2200, name: ""ao_table"": 93730"                <--- table
"adding OID assignment: catalog ""1247"", namespace: 2200, name: ""ao_table"": 93731"                <---composite type for table

"adding OID assignment: catalog ""1259"", namespace: 99, name: ""pg_toast_93730"": 93737"            <--- toast table
"adding OID assignment: catalog ""1247"", namespace: 99, name: ""pg_toast_93730"": 93738"            <--- composite type for toast table 
"adding OID assignment: catalog ""1259"", namespace: 99, name: ""pg_toast_93730_index"": 93739"      <--- index for toast table

"adding OID assignment: catalog ""1259"", namespace: 6104, name: ""pg_aoseg_93730"": 93732"          <--- pg_aoseg table
"adding OID assignment: catalog ""1247"", namespace: 6104, name: ""pg_aoseg_93730"": 93733"          <--- composite type for pg_aoseg table

"adding OID assignment: catalog ""1259"", namespace: 6104, name: ""pg_aoblkdir_93730"": 93740"       <--- blkdir table (created if `ao_table` has an index)
"adding OID assignment: catalog ""1247"", namespace: 6104, name: ""pg_aoblkdir_93730"": 93741"       <--- composite type for blkdir table
"adding OID assignment: catalog ""1259"", namespace: 6104, name: ""pg_aoblkdir_93730_index"": 93742" <--- index for blkdir table

"adding OID assignment: catalog ""1259"", namespace: 6104, name: ""pg_aovisimap_93730"": 93734"      <--- aovisimap table
"adding OID assignment: catalog ""1247"", namespace: 6104, name: ""pg_aovisimap_93730"": 93735"      <--- composite type for aovisimap table 
"adding OID assignment: catalog ""1259"", namespace: 6104, name: ""pg_aovisimap_93730_index"": 93736"<--- index for aovisimap table
```