@bmdoil 2022-07-27
> Reviewed most of this. Still digesting the last commit fully. I will give it another round later.
> 
> If possible, it would be really helpful if you could split the last commit into 4 separate commits - it would be much easier to review:
> 
> 1. Tackle AO
> 2. Tackle bitmap
> 3. Tackle types (changes to getTypes to get type composite/array type oids all at once)
> 4. Tackle toast
> 
> Mostly I looked at changes for 1 and 2.

I agree that it's a lot for a single commit. I spent quite awhile trying to split it into smaller chunks, but the changes are very interdependent and I wasn't able to split it cleanly. I can add a more thorough explanation of the design in the PR/Commit if it would help. Otherwise it might take a few days to unwind the changes and reapply incrementally.